+++ release_notes.md
@@ -10,3 +10,4 @@
 - Update PowerShell 7.4 worker to 4.0.4206
 - Re-add `Microsoft.DotNet.PlatformAbstractions` (#11031)
 - Update Python Worker Version to [4.37.0](https://github.com/Azure/azure-functions-python-worker/releases/tag/4.37.0)
+- Disable Diagnostic Events when Table Storage is not accessible (#10996)